requirements.txt
numpy
torch==2.5.1+cu121  # PyTorch with CUDA 12.1 support
torchvision==0.20.1+cu121
matplotlib
Pillow
pycocotools
opencv-python
ultralytics
